+++ takeout_watch/lib/app/bloc.dart
@@ -109,7 +109,11 @@ mixin AppBlocState {
       context.app.authenticated();
     }
     // prune incomplete/partial downloads
-    pruneCache(context.spiffCache.repository, context.trackCache.repository);
+    pruneCache(
+      spiffCache: context.spiffCache,
+      trackCache: context.trackCache,
+      settings: context.settings,
+    );
   }
 
   void appDispose() {}